requirements-conda.txt
blast=2.6.0
bmtagger=3.101
bwa=0.7.15
clark=1.2.3
diamond=0.8.36
gatk=3.6
kraken-all=0.10.6_eaf8fb68 # [linux]
krona=2.7
last=719
fastqc=0.11.5
mafft=7.221
mummer=3.23 
muscle=3.8.1551 
mvicuna=1.0
novoalign=3.07.00
picard=2.9.0
prinseq=0.20.4
samtools=1.3.1
snpeff=4.1l
tbl2asn=25.3
trimmomatic=0.36
trinity=date.2011_11_26
vphaser2=2.0
# Python packages below
biopython=1.68
matplotlib=1.5.3
future>=0.15.2
pysam=0.9.1
bedtools=2.26.0
pybedtools=0.7.8
PyYAML==3.11
pytest==3.0.5
pycodestyle
mock==2.0.0
six<2
pytest-cov==2.4.0
pytest-catchlog==1.2.2
pytest-xdist==1.15.0
flake8<=3
Jinja2==2.8
